gta2swifi: make setCameraLaunchStatus no-op
Change I2a12b5373e2ea851fa9d678c3cee9dab9daf25f8 was supposed to ensure
that setCameraLaunchStatus was never triggered, as it was "supposed" to
be guarded behind the service.bootanim.exit (or the resulting hex edit)
prop. If the prop was true, it would branch to said function.

That's great and all, but whatever Samsung did here, they evaluated
whether the prop was set to `0`, which, given the haxed prop doesn't
exist, will always be true.

At this point our only option is to hexedit the function to force it to
do nothing and return 0, which is android::OK.

Before patch:
```
setCameraLaunchStatus:
    push.w {r4, r5, r6, r7, r8, sb, sl, lr}  ; +0 = 2d e9 f0 47
    ldr    r5, [pc, #0xdc]                   ; +4 = 37 4d
```

After patch:
```
setCameraLaunchStatus:
    mov r0,#0
    bx lr  ; +0 = 4f f0 00 00 70 47
```

Change-Id: I01b6e6f63a447235928ab65a940ec38c7f93aa7d